% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_as_date}
\alias{dt_as_date}
\title{Convertir un vector de valores interpretables como epoch a dates}
\usage{
dt_as_date(epoch)
}
\arguments{
\item{epoch}{Un vector de valores interpretables como epoch}
}
\value{
Un vector de valores Date
}
\description{
Convertir un vector de valores interpretables como epoch a dates
}
