% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_as_posix}
\alias{dt_as_posix}
\title{Convertir un vector de valores interpretables como epoch a POSIXct}
\usage{
dt_as_posix(epoch)
}
\arguments{
\item{epoch}{Un vector de valores interpretables como epoch}
}
\value{
Un vector de valores POSIXct
}
\description{
Convertir un vector de valores interpretables como epoch a POSIXct
}
