% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_as_tms}
\alias{dt_as_tms}
\title{Convertir un vector de valores interpretables como epoch a cadenas de Timestamp}
\usage{
dt_as_tms(epoch)
}
\arguments{
\item{epoch}{Un vector de valores interpretables como epoch}
}
\value{
Un vector de cadenas con formato Timestamp
}
\description{
Convertir un vector de valores interpretables como epoch a cadenas de Timestamp
}
