% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_as_unix}
\alias{dt_as_unix}
\title{Convertir fechas a formato numerico Unix}
\usage{
dt_as_unix(date)
}
\arguments{
\item{date}{Valor de tipo fecha a convertir}
}
\value{
Numero interpretable como un epoch Unix
}
\description{
Convertir fechas a formato numerico Unix
}
