% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_elapsed}
\alias{dt_elapsed}
\title{Calcula y formatea un intervalo de tiempo}
\usage{
dt_elapsed(start, end)
}
\arguments{
\item{start}{Inicio del intervalo}

\item{end}{Fin del intervalo}
}
\value{
El tiempo consumido formateado
}
\description{
Calcula y formatea un intervalo de tiempo
}
