% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{dt_round2Posix}
\alias{dt_round2Posix}
\title{Redondea la fecha al intervalo dado}
\usage{
dt_round2Posix(date, interval, prev = TRUE)
}
\arguments{
\item{date}{Fecha a redondear}

\item{interval}{Intervalo de redondeo}

\item{prev}{Redondear a la fecha anterior/siguiente o igual. Defecto: Anterior}
}
\value{
La fecha redondeada en formato POSIXct
}
\description{
Redondea la fecha al intervalo dado
}
