% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_title_case}
\alias{str_title_case}
\title{Convertir cadenas a modo titulo}
\usage{
str_title_case(texts, locale = "es")
}
\arguments{
\item{texts}{Un vector de caracteres}

\item{locale}{Codigo de idioma a usar. Por defecto "es"}
}
\value{
Un vector de caracteres de la misma longitud que \code{texts} en modo titulo
}
\description{
Convertir cadenas a modo titulo
}
