% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_to_lower}
\alias{str_to_lower}
\title{Convertir las cadenas a minusculas}
\usage{
str_to_lower(texts)
}
\arguments{
\item{texts}{Un vector de caracteres}
}
\value{
Un vector de caracteres de la misma longitud que \code{texts} con los caracteres en minusculas
}
\description{
Convertir las cadenas a minusculas
}
