% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_to_upper}
\alias{str_to_upper}
\title{Convertir las cadenas a mayusculas}
\usage{
str_to_upper(texts)
}
\arguments{
\item{texts}{Un vector de caracteres}
}
\value{
Un vector de caracteres de la misma longitud que \code{texts} con los caracteres en mayusculas
}
\description{
Convertir las cadenas a mayusculas
}
